requirements.txt
structlog
streamlit
numpy
pandas
mysql-connector-python
leafmap
geopandas
plotly